emoji-search/requirements.txt
streamlit==1.35.0
sentence-transformers==2.6.1
torch==2.2.2
google-generativeai==0.5.2
python-dotenv==1.0.1
